+++ src/config.rs
@@ -67,6 +67,8 @@ pub struct Config {
     pub no_answer: bool,
     #[serde(default = "default_false")]
     pub use_delimiters: bool,
+    #[serde(default = "default_false")]
+    pub local_run: bool,
 }
 
 impl Config {
@@ -85,6 +87,7 @@ impl Config {
             lifetime: default_lifetime(),
             no_answer: default_true(),
             use_delimiters: default_false(),
+            local_run: default_false(),
         }
     }
 
@@ -110,6 +113,7 @@ pub fn clear_config(
     lifetime: bool,
     no_answer: bool,
     use_delimiters: bool,
+    local_run: bool,
 ) -> Result<(), String> {
     if url {
         conf.url = default_url();
@@ -147,9 +151,12 @@ pub fn clear_config(
     if use_delimiters {
         conf.use_delimiters = default_false();
     }
+    if local_run {
+        conf.local_run = default_false();
+    }
 
     if (url || addr || wallet || abi || keys || retries || timeout || wc || depool_fee || lifetime
-        || no_answer || use_delimiters) == false {
+        || no_answer || use_delimiters || local_run) == false {
         conf = Config::new();
     }
     let conf_str = serde_json::to_string(&conf)
@@ -175,6 +182,7 @@ pub fn set_config(
     lifetime:  Option<&str>,
     no_answer:  Option<&str>,
     use_delimiters: Option<&str>,
+    local_run: Option<&str>,
 ) -> Result<(), String> {
     if let Some(s) = url {
         conf.url = s.to_string();
@@ -221,6 +229,10 @@ pub fn set_config(
     if let Some(use_delimiters) = use_delimiters {
         conf.use_delimiters = use_delimiters.parse::<bool>()
             .map_err(|e| format!(r#"failed to parse "use_delimiters": {}"#, e))?;
+    }
+    if let Some(local_run) = local_run {
+        conf.local_run = local_run.parse::<bool>()
+            .map_err(|e| format!(r#"failed to parse "local_run": {}"#, e))?;
     }
 
     let conf_str = serde_json::to_string(&conf)