+++ CHANGELOG.md
@@ -60,6 +60,6 @@ This fixes affects all debots invoking other debots (e.g. depool debot, mludi de
 ## v0.12.0
 
 ### New feature
-- tonos-cli now executes transaction locally for deploy and call commands before executing it 
-  onchain. If local execution fails, onchain execution is not performed. Local run can be
-  cancelled by enabling flag `no_local_run` in the tonos-cli config. 
+- tonos-cli now can execute the transaction locally for deploy and call commands before executing
+  it onchain. If local execution fails, onchain execution is not performed. Local run can be
+  enabled by setting the flag `local_run` in the tonos-cli config.